# Compare performance of StringBuilder and String concatenation