--- a/src/StringBuilderTime.java
+++ b/src/StringBuilderTime.java
@@ -1,17 +1,14 @@
 class StringBuilderTime {
 
-    String buildString(int length) {
+    public static void main(String[] args) {
+        System.out.println("Enter length: " + args[0]);
+        int length = Integer.parseInt(args[0]);
+
         StringBuilder result = new StringBuilder();
         for (int i = 0; i < length; i++) {
             result.append("D");
         }
-        return "Finished creating a string consisting of " + length + " characters.";
-    }
 
-    public static void main(String[] args) {
-        System.out.println("Enter length: " + args[0]);
-
-        StringBuilderTime stringBuilderTime = new StringBuilderTime();
-        System.out.println(stringBuilderTime.buildString(Integer.parseInt(args[0])));
+        System.out.println("Finished creating a string consisting of " + length + " characters.");
     }
 }