--- a/src/StringBuilderTime.java
+++ b/src/StringBuilderTime.java
@@ -1,7 +1,6 @@
 class StringBuilderTime {
 
     public static void main(String[] args) {
-        System.out.println("Enter length: " + args[0]);
         int length = Integer.parseInt(args[0]);
 
         StringBuilder result = new StringBuilder();