--- a/src/StringTime.java
+++ b/src/StringTime.java
@@ -1,6 +1,6 @@
 class StringTime {
 
-    String printer(int length) {
+    String buildString(int length) {
         String result = "";
         for (int i = 0; i < length; i++) {
             result = result + "D";
@@ -12,6 +12,6 @@ class StringTime {
         System.out.println("Enter length: " + args[0]);
 
         StringTime stringTime = new StringTime();
-        System.out.println(stringTime.printer(Integer.parseInt(args[0])));
+        System.out.println(stringTime.buildString(Integer.parseInt(args[0])));
     }
 }