**Result of comparing performance of StringBuilder and String concatenation with Time command:**

```
$ time java StringTime.java 10000
Finished creating a string consisting of 10000 characters.
java StringTime.java 10000  0.52s user 0.06s system 126% cpu 0.466 total
```
Test StringBuilder performance with:

- When the string length is 10,000: 0.55s user, 0.07s system, 128% CPU, 0.480 total.
- When the string length is 100,000: 0.55s user, 0.06s system, 130% CPU, 0.467 total.
-  When the string length is 1,000,000: 0.57s user, 0.04s system, 187% CPU, 0.325 total.

Test String concatenation performance:

- When the string length is 10,000: 0.54s user, 0.04s system, 177% CPU, 0.324 total.
- When the string length is 100,000: 0.99s user, 0.16s system, 142% CPU, 0.804 total.
-  When the string length is 1,000,000: 23.81s user, 0.50s system, 111% CPU, 21.886 total.